+++ src/joomart/Joomart.java
@@ -1,4 +1,4 @@
-package JoomartsPack;
+package joomart;
 
 public class Joomart {
     public static void main(String[] args) {
