+++ src/joomart/NewJoomartsClass.java
@@ -1,0 +1,7 @@
+package joomart;
+
+public class NewJoomartsClass {
+    public static void main(String[] args) {
+        System.out.println("git practice");
+    }
+}
